+++ src/Servidor/Conexion/MovementMessageSerializer.h
@@ -8,7 +8,7 @@
 class MovementMessageSerializer: public Serializer {
 public:
     std::string serialize(Message* messages) override;
-    Message* deserialize(unsigned char * data,uint16_t len_data) override;
+    Message* deserialize(unsigned char * data,uint32_t len_data) override;
 };
