+++ src/Servidor/Conexion/SpawnPcMessageSerializer.cpp
@@ -9,7 +9,7 @@ std::string SpawnPcMessageSerializer::serialize(Message *message) {
     return result;
 }
 
-Message *SpawnPcMessageSerializer::deserialize(unsigned char *data, uint16_t len_data) {
+Message *SpawnPcMessageSerializer::deserialize(unsigned char *data, uint32_t len_data) {
     msgpack::object_handle oh =
             msgpack::unpack(reinterpret_cast<const char *>(data), len_data);
