+++ src/Servidor/Conexion/SpawnStaticObjectMessageSerializer.cpp
@@ -10,7 +10,7 @@ std::string SpawnStaticObjectMessageSerializer::serialize(Message *message) {
     return result;
 }
 
-Message * SpawnStaticObjectMessageSerializer::deserialize(unsigned char *data,uint16_t len_data) {
+Message * SpawnStaticObjectMessageSerializer::deserialize(unsigned char *data,uint32_t len_data) {
     msgpack::object_handle oh =
             msgpack::unpack(reinterpret_cast<const char *>(data), len_data);
